requirements.txt
asgiref==3.10.0
channels==4.3.1
channels_redis==4.3.0
contourpy==1.3.3
cycler==0.12.1
Django==5.2.8
django-admin-sortable2==2.2.8
django-filter==25.2
django-tailwind==4.4.1
djangorestframework==3.16.1
fonttools==4.60.1
gunicorn==23.0.0
kiwisolver==1.4.9
matplotlib==3.10.7
msgpack==1.1.2
numpy==2.3.4
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pyparsing==3.2.5
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
redis==7.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
